chore(test): 스텝별 테스트 패키지로 묶음 처리
