+++ src/test/java/roomescape/missionstep/MissionStep2Test.java
@@ -1,10 +1,11 @@
-package roomescape;
+package roomescape.missionstep;
 
 import io.restassured.RestAssured;
 import io.restassured.http.ContentType;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.jdbc.core.JdbcTemplate;
+import roomescape.ClearDbTest;
 import roomescape.domain.Reservation;
 
 import java.sql.Connection;
