+++ src/test/java/roomescape/missionstep/MissionStep4Test.java
@@ -1,8 +1,9 @@
-package roomescape;
+package roomescape.missionstep;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.jdbc.core.JdbcTemplate;
+import roomescape.ClearDbTest;
 import roomescape.controller.ReservationController;
 
 import java.lang.reflect.Field;
